Move inhibitors to own file
